Add a crude shim for clock_gettime on older macos:
`clock_gettime` has only been available on mac since 10.12 (Sierra), so shim it on older vers.

This shim is lower resolution, but that matters a lot less given that this usage only wants CLOCK_REALTIME_COARSE. It may not have precisely the same semantics about consistency, but I'm not sure how real a problem that is for this use case.