Add resource stats for macos:
The timings here are a little cruder than the linux versions, but as with the clock_gettime shim, I'm not sure they're so far from reality as to be problematic.

The I/O stats, however, are much worse. AFAICT macos has no way to get io counters equivalent to what procfs provides. The closest thing I could find would be dtrace probes, which would be ridiculously unwieldy (and I expect costly) for this use case, not to mention requiring root and a one-off SIP reconfig.

I've opted for basic disk io stats from getrusage, based on (hardcoded) block size. Of course, this doesn't include network io, which is a major difference to the linux version.