Make a test not affected by libc implementation
`spx_013.phpt` test calls `foo` function 36150 times. This value is
then converted to `36.15` and then formatted by printf with `%7.1f`
as format.

The expected formatted value is `36.1`, which is the output of glibc.
However musl libc output is `36.2`. This is due to the fact that
musl libc *printf implementation takes default FP rounding mode into
account (round half to even) while glibc does not.

This fix is just a workaround to make this test not affected by this
difference between these 2 libc implementations.

Fixes #84