Sampling profiler: improve performance
The current implementation of sampling profiler decorator, based on
XHProf one, has still a lower (compared to tracing profiler) but
significant overhead due to the fact that it calls `clock_gettime()`
on each function call/start.

This patch make it use a background thread.

It may also fix #91 as `last_sample_time_ns` was not properly
initialized in previous implementation.