Allow to define the profiled spans at runtime
This patch adds the support for self-code instrumentation in order
to let the user to trigger the profiling for a specific part of
the script at runtime.

The 2 following userland functions are added:
- `spx_profiler_start(): void`
- `spx_profiler_stop(): void`

This feature requires the PHP script to be configured with:
- SPX_ENABLED=1
- SPX_AUTO_START=0 -> this is a newly added parameter (with `1` as
    default value) for the purpose of this feature.

See the updated documentation for more details.

Resolves #107