Reduce "full" report mode overhead
This patch uses Zstandard compression for "full" report data file.
Zstandard is faster than zlib (DEFLATE) while achieving a better
compression ratio.

For instance a 208MB report is compressed to:
- 49MB with zlib / level 6.
- 43MB with Zstandard / level 1.

Zstandard is also OK for the web UI since it is supported by both
Firefox and Chromium since few years.

The per-function call overhead is thus significantly reduced, going
from ~427ns to ~273ns.

This patch allows SPX to match the peformance of Xhprof.

As a reminder:
- timings have been measured on my hardware (AMD Ryzen 7 9700X)
    and with my php-profiler-bench project.
- SPX v0.4's (v0.4.18) per-function call overhead is 565ns.
- current SPX's per-function call overhead is 273ns.
- Xhprof's (v2.3.10) per-function call overhead is 274ns.
- Xdebug's (v3.4.1) per-function call overead is 1595ns.